Replacing colorize with ANSI
Unfortunately, the colorize gem's API is somewhat unstable. The gem itself also
happens to be licensed under the GPLv2.

While I applaud the GNU's idealism, I refuse to impose the GPL's obligations on
the users and distributors of middleman-s3_sync. I am now using the ANSI library
to colorize the output of middleman-s3_sync. It is licensed under the BSD
license.

This also happens to fix #70.