Updating dependencies.
This addresses #110 where fog's removal of the mime-types library
caused a failure when trying to upload content to S3. The workaround
is to add the mime-types gem to a middleman's Gemfile. I am adding this
to middleman-s3_sync since it is likely to affect most users.

The issue has been reported to Fog since it is likely to impact other
gems depending on it.